requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
decorator==5.1.1
fonttools==4.42.1
future==0.18.3
geocoder==1.38.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tk==0.1.0
tkintermapview==1.29
urllib3==2.0.4
